# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alive — Ghost Worker Verification API",
        "description": "AI proof-of-life verification system for Nigerian government payroll.",
        "contact": {
            "email": "api@sentinelbloom.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/alerts": {
            "get": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "List ghost worker alerts",
                "operationId": "alertIndex",
                "parameters": [
                    {
                        "name": "severity",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "critical"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mda_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated alerts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}": {
            "get": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Get a single alert with dispatch history",
                "operationId": "alertShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/block": {
            "post": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Block salary for a ghost alert",
                "operationId": "alertBlock",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Salary blocked"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/dispatch": {
            "post": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Dispatch nearest field agent to investigate",
                "operationId": "alertDispatch",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent dispatched"
                    },
                    "422": {
                        "description": "No available agent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/refer-icpc": {
            "post": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Refer an alert to the ICPC",
                "operationId": "alertReferIcpc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert referred to ICPC"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/false-positive": {
            "post": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Mark an alert as a false positive",
                "operationId": "alertFalsePositive",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alert marked as false positive"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/resolve": {
            "post": {
                "tags": [
                    "Ghost Worker Alerts"
                ],
                "summary": "Resolve a ghost worker alert",
                "operationId": "alertResolve",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "notes": {
                                        "type": "string",
                                        "example": "Worker physically verified at duty post"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alert resolved"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}